refactor geometry package. add geometry demonstration notebook. refactor, correct, and complete function and map system base classes, cylinder, and torus.
