+++ pyzeta/core/dynamics/function_systems/implementations/hyperbolic_cylinder.py
@@ -6,7 +6,7 @@ Authors:\n
 - Sebastian Albrecht\n
 """
 
-from typing import Tuple
+from typing import List, Tuple
 
 import numpy as np
 
@@ -17,19 +17,14 @@ from pyzeta.core.dynamics.function_systems.moebius_system import (
 from pyzeta.core.dynamics.function_systems.schottky_exception import (
     InvalidSchottkyException,
 )
-from pyzeta.core.dynamics.function_systems.schottky_surface import (
-    SchottkySurface,
+from pyzeta.core.dynamics.function_systems.schottky_system import (
+    SchottkyFunctionSystem,
+    SchottkyMapSystem,
 )
-from pyzeta.core.pyzeta_types.general import (
-    tBoolMat,
-    tIndexVec,
-    tMatVec,
-    tVec,
-    tWordVec,
-)
+from pyzeta.core.pyzeta_types.general import tMatVec
 
 
-class HyperbolicCylinder(SchottkySurface):
+class HyperbolicCylinder(SchottkyFunctionSystem):
     """
     Class representation a hyperbolic cylinder, i.e. the single topological
     possibility for a Schottky surface of rank one.
@@ -37,7 +32,7 @@ class HyperbolicCylinder(SchottkySurface):
 
     __slots__ = ("width",)
 
-    def __init__(self, funnelWidth: float) -> None:
+    def __init__(self, funnelWidth: float, rotate: bool) -> None:
         """
         Initialize a hyperbolic cylinder.
 
@@ -46,10 +41,17 @@ class HyperbolicCylinder(SchottkySurface):
         closed geodesic).
 
         :param funnelWidth: width of the cylinder at its most narrow point
+        :param rotate: flag indicating whether to rotate cylinder
         """
         self.width = funnelWidth
         self.logger.info("creating %s", str(self))
-        gen = [[np.exp(self.width / 2), 0], [0, np.exp(-self.width / 2)]]
+        if not rotate:
+            gen = [[np.exp(self.width / 2), 0], [0, np.exp(-self.width / 2)]]
+        else:
+            gen = [
+                [np.cosh(self.width / 2), np.sinh(self.width / 2)],
+                [np.sinh(self.width / 2), np.cosh(self.width / 2)],
+            ]
 
         try:
             super().__init__(np.array([gen], dtype=np.float64))
@@ -61,44 +63,20 @@ class HyperbolicCylinder(SchottkySurface):
         return f"HyperbolicCylinder({self.width:.4g})"
 
 
-class HyperbolicCylinderMap(MoebiusMapSystem):
+class HyperbolicCylinderMap(SchottkyMapSystem):
     "Class representing hyperbolic cylinders as hyperbolic map systems."
-
-    __slots__ = ("cylinderSystem",)
 
     def __init__(self, funnelWidth: float) -> None:
         """
-        Initialize a hyperbolic map system representation of the hyperbolic
-        cylinder. This map system delegates most work to an associated function
-        system.
+        Initialize a Moebius hyperbolic map system representation of the
+        hyperbolic cylinder. The underlying flow adapted cylinder is always
+        rotated to obtain sensible fundamental intervals.
 
         :param funnelWidth: width of the cylinder at its most narrow point
         """
-        self.cylinderSystem = HyperbolicCylinder(funnelWidth=funnelWidth)
-
-    # docstr-coverage: inherited
-    @property
-    def fundamentalIntervals(self) -> Tuple[Tuple[float, float], ...]:
-        raise NotImplementedError()
-
-    # TODO: put this into abstract base class!
-    # docstr-coverage: inherited
-    @property
-    def adjacencyMatrix(self) -> tBoolMat:
-        return self.cylinderSystem.adjacencyMatrix
-
-    # TODO: put this into abstract base class!
-    # docstr-coverage: inherited
-    def getStabilities(self, words: tWordVec) -> Tuple[tVec, tVec]:
-        contractingStabs = self.cylinderSystem.getStabilities(words)
-        return contractingStabs, 1 / contractingStabs
-
-    # docstr-coverage: inherited
-    def getPeriodicPoints(self, words: tWordVec) -> Tuple[tVec, tVec]:
-        raise NotImplementedError()
-
-    def getGenerators(self, indices: tIndexVec) -> tMatVec:
-        raise NotImplementedError()
+        super().__init__(
+            HyperbolicCylinder(funnelWidth=funnelWidth, rotate=True)
+        )
 
 
 class FlowAdaptedCylinder(MoebiusFunctionSystem):
@@ -144,12 +122,12 @@ class FlowAdaptedCylinder(MoebiusFunctionSystem):
             gen2 = [[0, np.exp(-self.width / 4)], [np.exp(self.width / 4), 0]]
         else:
             gen1 = [
-                [np.cosh(self.width / 4), -np.sinh(self.width / 4)],
-                [np.sinh(self.width / 4), -np.cosh(self.width / 4)],
-            ]
-            gen2 = [
                 [np.cosh(self.width / 4), np.sinh(self.width / 4)],
                 [-np.sinh(self.width / 4), -np.cosh(self.width / 4)],
+            ]
+            gen2 = [
+                [np.cosh(self.width / 4), -np.sinh(self.width / 4)],
+                [np.sinh(self.width / 4), -np.cosh(self.width / 4)],
             ]
         generators = [gen1, gen2, gen1, gen2]
         return np.array(generators, dtype=np.float64)
@@ -157,3 +135,56 @@ class FlowAdaptedCylinder(MoebiusFunctionSystem):
 
 class FlowAdaptedCylinderMap(MoebiusMapSystem):
     "Class representation of flow adapted cylinders as hyperbolic map systems."
+
+    __slots__ = ("_fundInter",)
+
+    def __init__(self, funnelWidth: float) -> None:
+        """
+        Initialize a Moebius hyperbolic map system representation of the flow
+        adapted (hyperbolic) cylinder. The underlying flow adapted cylinder is
+        always rotated to obtain sensible fundamental intervals.
+
+        :param funnelWidth: width of the cylinder at its most narrow point
+        """
+        super().__init__(FlowAdaptedCylinder(funnelWidth, rotate=True))
+
+        self._fundInter, deltaOffset = self._getFundamentalIntervals()
+        self._addTranslationToReflections(deltaOffset)
+
+    def _getFundamentalIntervals(
+        self,
+    ) -> Tuple[Tuple[Tuple[float, float], ...], float]:
+        """
+        TODO.
+        """
+        firstHalf: List[Tuple[float, float]] = []
+        for reflection in self._functionSystem._phi[:2]:
+            radius = 1.0 / reflection[1, 0]
+            midPoint = reflection[0, 0] * radius
+            firstHalf.append((midPoint - abs(radius), midPoint + abs(radius)))
+        deltaOffset = firstHalf[1][1] - firstHalf[0][0] + 1
+        secondHalf: List[Tuple[float, float]] = []
+        for interval in firstHalf:
+            secondHalf.append(
+                (interval[0] + deltaOffset, interval[1] + deltaOffset)
+            )
+
+        return tuple(firstHalf + secondHalf), deltaOffset
+
+    def _addTranslationToReflections(self, deltaOffset: float) -> None:
+        """
+        TODO.
+        """
+        translation = np.array([[1, deltaOffset], [0, 1]], dtype=np.float64)
+        translationInv = np.array(
+            [[1, -deltaOffset], [0, 1]], dtype=np.float64
+        )
+        for i, reflection in enumerate(self._functionSystem._phi[:2]):
+            self._functionSystem._phi[i] = translation @ reflection
+        for i, reflection in enumerate(self._functionSystem._phi[2:]):
+            self._functionSystem._phi[i + 2] = reflection @ translationInv
+
+    # docstr-coverage: inherited
+    @property
+    def fundamentalIntervals(self) -> Tuple[Tuple[float, float], ...]:
+        return self._fundInter
